pyproject.toml
[project]
name = "quadre"
version = "0.1.1"
description = "A Python CLI tool that generates beautiful dashboard images from JSON data using Cairo"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
 "pycairo>=1.28.0",
 "pydantic>=2.7,<3",
]
classifiers = [ "Typing :: Typed", "License :: OSI Approved :: MIT License" ]


[project.urls]
Repository = "https://github.com/amrltqt/quadre"
Issues = "https://github.com/amrltqt/quadre/issues"


[[project.authors]]
name = "Quentin Amirault"
email = "amirault.quentin@gmail.com"


[project.license]
file = "LICENSE"


[project.scripts]
quadre = "quadre.cli:main"
quadre-validate = "quadre.cli:main_validate"


[build-system]
requires = [ "hatchling" ]
build-backend = "hatchling.build"


[tool.hatch.build.targets.wheel]
packages = [ "src/quadre" ]
include = [ "src/quadre/theme/*.json", "src/quadre/py.typed", "LICENSE" ]


[tool.hatch.build.targets.sdist]
include = [
"src/quadre",
"src/quadre/theme/*.json",
"src/quadre/py.typed",
"README.md",
"LICENSE"
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.pyright]
include = ["src"]
pythonPath = "src"
typeCheckingMode = "basic"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.11",
    "twine>=6.1.0"
]
